[#3] factory 패턴 부분 구현
